Corrections to the drop mechanic of the scheme pane. The scheme pane's remove button now works for SIPs too. Removed the "remove" button from the inspection pane.
